simplify ReverseRequest
Merge ReverseRequest classes into a single class and get rid of the
factory, which only ever returned the same object.
